[Refactor]: Collocate markets constants inside crypto folder under components
