+++ src/components/Crypto/constants/index.ts
